refactored entire application
