--- bin/state-handlers/show-error.js
@@ -1,3 +1,0 @@
-export default async (stateMachine) => {
-    process.stderr.write(stateMachine.context.error.message + '\n')
-}
